Add files via upload
Steganography Project
This project is a web application for performing basic steganography and encryption tasks. It includes functionality for encoding and decoding messages hidden in images and provides encryption and decryption of text using a custom affine cipher. The application is built with HTML, CSS, and JavaScript.

Features
Affine Cipher Encryption/Decryption: Encrypt and decrypt messages using an affine cipher with customizable parameters a and b.
Message Encoding: Hide a text message within an image using steganography techniques.
Message Decoding: Extract and display the hidden message from an encoded image.
User-Friendly Interface: Simple and intuitive user interface for interacting with the various features.
Technologies Used
HTML
CSS
JavaScript
Usage
Affine Cipher:

Enter a message, and the values for a and b.
Click "Encrypt" to encode the message.
Click "Decrypt" to decode the message.
Message Encoding:

Enter the message to be hidden.
Select an image file.
Click "Encode" to hide the message within the image and download the encoded image.
Message Decoding:

Select an image file containing a hidden message.
The hidden message will be extracted and displayed.
How to Run
Open the index.html file in a web browser to start using the application.

Contribution
Feel free to fork this repository, make improvements, and submit pull requests. Any contributions to enhance the functionality or UI are welcome!